ai_resume_backend/requirements.txt
fastapi
uvicorn
jinja2
xhtml2pdf
python-dotenv
requests
pydantic

